+++ apps/api/src/modules/test-runs/services/test-runs-mapper.service.ts
@@ -55,6 +55,7 @@ export class TestRunsMapperService {
       expired: entity.expired,
       valid: entity.valid,
       reasons_not_valid: entity.reasonsNotValid,
+      data_warnings: entity.dataWarnings,
       adapt_config: entity.adaptConfig,
       variables: entity.variables,
       deep_links: entity.deepLinks,